refactor(presentation): adopt atomic design structure\n\n- Extract atoms: InfoChip\n- Extract molecules: PlaybackControls, TrailControls, DoubleParamsForm (moved)\n- Extract organisms: DoublePendulumAnimator (moved)\n- Update HomePage to compose atomic components\n\nNo functional changes intended.
