+++ .gitignore
@@ -43,3 +43,8 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+
+# Agents
+AGENTS.md
+GEMINI.md
+GPT.md
