+++ lib/presentation/atoms/info_chip.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class InfoChip extends StatelessWidget {
+  final String label;
+  final String value;
+  const InfoChip({super.key, required this.label, required this.value});
+
+  @override
+  Widget build(BuildContext context) {
+    return Chip(
+      label: Text('$label: $value'),
+      backgroundColor: Theme.of(context).colorScheme.primary.withOpacity(0.08),
+      side: BorderSide.none,
+    );
+  }
+}
+
