+++ lib/presentation/pages/home_page.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:rk4_solver/application/double_pendulum_controller.dart';
-import 'package:rk4_solver/presentation/widgets/double_params_form.dart';
-import 'package:rk4_solver/presentation/widgets/double_pendulum_animator.dart';
+import 'package:rk4_solver/presentation/molecules/playback_controls.dart';
+import 'package:rk4_solver/presentation/molecules/trail_controls.dart';
+import 'package:rk4_solver/presentation/atoms/info_chip.dart';
+import 'package:rk4_solver/presentation/molecules/double_params_form.dart';
+import 'package:rk4_solver/presentation/organisms/double_pendulum_animator.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -72,9 +75,9 @@ class _HomePageState extends State<HomePage> {
                 spacing: 8,
                 runSpacing: 4,
                 children: [
-                  _InfoChip(label: '計算点数', value: result.points.length.toString()),
-                  _InfoChip(label: 'シミュレーション時間 [秒]', value: _controller.params.tEnd.toStringAsFixed(2)),
-                  _InfoChip(label: '時間刻み [秒]', value: _controller.params.h.toStringAsFixed(4)),
+                  InfoChip(label: '計算点数', value: result.points.length.toString()),
+                  InfoChip(label: 'シミュレーション時間 [秒]', value: _controller.params.tEnd.toStringAsFixed(2)),
+                  InfoChip(label: '時間刻み [秒]', value: _controller.params.h.toStringAsFixed(4)),
                 ],
               ),
             ],
@@ -103,14 +106,14 @@ class _HomePageState extends State<HomePage> {
               ),
             ),
             const SizedBox(height: 8),
-            _PlaybackControls(
+            PlaybackControls(
               playing: _playing,
               speed: _speed,
               onToggle: () => setState(() => _playing = !_playing),
               onSpeedChanged: (v) => setState(() => _speed = v),
             ),
             const SizedBox(height: 8),
-            _TrailControls(
+            TrailControls(
               showTrail: _showTrail,
               trailCount: _trailCount,
               onShowTrailChanged: (v) => setState(() => _showTrail = v),
@@ -122,126 +125,4 @@ class _HomePageState extends State<HomePage> {
     );
   }
 }
-
-class _InfoChip extends StatelessWidget {
-  final String label;
-  final String value;
-  const _InfoChip({required this.label, required this.value});
-
-  @override
-  Widget build(BuildContext context) {
-    return Chip(
-      label: Text('$label: $value'),
-      backgroundColor: Theme.of(context).colorScheme.primary.withOpacity(0.08),
-      side: BorderSide.none,
-    );
-  }
-}
-
-class _PlaybackControls extends StatelessWidget {
-  final bool playing;
-  final double speed;
-  final VoidCallback onToggle;
-  final ValueChanged<double> onSpeedChanged;
-  const _PlaybackControls({
-    required this.playing,
-    required this.speed,
-    required this.onToggle,
-    required this.onSpeedChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Row(
-              children: [
-                IconButton(
-                  iconSize: 28,
-                  onPressed: onToggle,
-                  icon: Icon(playing ? Icons.pause_circle_filled : Icons.play_circle_fill),
-                  color: Theme.of(context).colorScheme.primary,
-                ),
-                const SizedBox(width: 8),
-                Text(
-                  playing ? '再生中' : '一時停止',
-                  style: Theme.of(context).textTheme.titleMedium,
-                ),
-                const Spacer(),
-                Text('${speed.toStringAsFixed(2)}x'),
-              ],
-            ),
-            Row(
-              children: [
-                const Text('速度'),
-                Expanded(
-                  child: Slider(
-                    value: speed,
-                    min: 0.5,
-                    max: 2.0,
-                    divisions: 15,
-                    label: '${speed.toStringAsFixed(2)}x',
-                    onChanged: onSpeedChanged,
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _TrailControls extends StatelessWidget {
-  final bool showTrail;
-  final int trailCount;
-  final ValueChanged<bool> onShowTrailChanged;
-  final ValueChanged<double> onTrailCountChanged;
-  const _TrailControls({
-    required this.showTrail,
-    required this.trailCount,
-    required this.onShowTrailChanged,
-    required this.onTrailCountChanged,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            SwitchListTile(
-              contentPadding: EdgeInsets.zero,
-              title: const Text('軌跡を表示'),
-              value: showTrail,
-              onChanged: onShowTrailChanged,
-            ),
-            if (showTrail)
-              Row(
-                children: [
-                  const Text('軌跡長'),
-                  Expanded(
-                    child: Slider(
-                      value: trailCount.toDouble().clamp(20, 500),
-                      min: 20,
-                      max: 500,
-                      divisions: 24,
-                      label: trailCount.toString(),
-                      onChanged: onTrailCountChanged,
-                    ),
-                  ),
-                ],
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-}
+
